// migrations/snapshots/644f1de58ee9025d8dc07ce21ad3ca86f0c7cb8e339f657d7325c3509bf902c5/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Review: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly targetType: 'TRAINER' | 'CENTER' | 'SESSION' | 'BOOKING' | 'PAYMENT';
      readonly targetId: CodecTypes['pg/text@1']['output'];
      readonly rating: CodecTypes['pg/int4@1']['output'] | null;
      readonly comment: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Streak: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly currentStreak: CodecTypes['pg/int4@1']['output'];
      readonly longestStreak: CodecTypes['pg/int4@1']['output'];
      readonly lastActivityDate: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly phone: CodecTypes['pg/int4@1']['output'] | null;
      readonly role: 'USER' | 'ADMIN';
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly profileImage: CodecTypes['pg/text@1']['output'] | null;
      readonly city: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Review: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly targetType: 'TRAINER' | 'CENTER' | 'SESSION' | 'BOOKING' | 'PAYMENT';
      readonly targetId: CodecTypes['pg/text@1']['input'];
      readonly rating: CodecTypes['pg/int4@1']['input'] | null;
      readonly comment: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Streak: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly currentStreak: CodecTypes['pg/int4@1']['input'];
      readonly longestStreak: CodecTypes['pg/int4@1']['input'];
      readonly lastActivityDate: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly phone: CodecTypes['pg/int4@1']['input'] | null;
      readonly role: 'USER' | 'ADMIN';
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly profileImage: CodecTypes['pg/text@1']['input'] | null;
      readonly city: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly review: {
      readonly comment: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly rating: CodecTypes['pg/int4@1']['output'] | null;
      readonly targetId: CodecTypes['pg/text@1']['output'];
      readonly targetType: 'TRAINER' | 'CENTER' | 'SESSION' | 'BOOKING' | 'PAYMENT';
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly streak: {
      readonly currentStreak: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly lastActivityDate: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly longestStreak: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly city: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly phone: CodecTypes['pg/int4@1']['output'] | null;
      readonly profileImage: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'USER' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly review: {
      readonly comment: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly rating: CodecTypes['pg/int4@1']['input'] | null;
      readonly targetId: CodecTypes['pg/text@1']['input'];
      readonly targetType: 'TRAINER' | 'CENTER' | 'SESSION' | 'BOOKING' | 'PAYMENT';
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly streak: {
      readonly currentStreak: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly lastActivityDate: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly longestStreak: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly city: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly phone: CodecTypes['pg/int4@1']['input'] | null;
      readonly profileImage: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'USER' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;